fix: move components to correct directory
